refactor(api): move course repository to src
