+++ api/tests/shared/integration/infrastructure/repositories/course-repository_test.js
@@ -1,7 +1,7 @@
-import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { Course } from '../../../../lib/domain/models/Course.js';
-import * as courseRepository from '../../../../lib/infrastructure/repositories/course-repository.js';
-import { catchErr, domainBuilder, expect, mockLearningContent } from '../../../test-helper.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { Course } from '../../../../../lib/domain/models/Course.js';
+import * as courseRepository from '../../../../../src/shared/infrastructure/repositories/course-repository.js';
+import { catchErr, domainBuilder, expect, mockLearningContent } from '../../../../test-helper.js';
 
 describe('Integration | Repository | course-repository', function () {
   describe('#get', function () {
